backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
